requirements.txt
# Requirements.txt cannot contain Korean characters. (results in cp949 error)
# Never use Korean characters.

pyinstaller
pyinstaller-hooks-contrib
python-dotenv
dduk-core
dduk-process
flask